backend/app/utils/security.py
from datetime import datetime, timedelta
from typing import Any, Dict
import jwt
from passlib.context import CryptContext
from ..config import Settings

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")


def create_access_token(subject: int, expires_minutes: int | None = None) -> str:
    expire = datetime.utcnow() + timedelta(minutes=expires_minutes or Settings.access_token_expire_minutes)
    to_encode: Dict[str, Any] = {"sub": str(subject), "exp": expire}
    return jwt.encode(to_encode, Settings.jwt_secret_key, algorithm=Settings.jwt_algorithm)


def decode_access_token(token: str) -> Dict[str, Any]:
    return jwt.decode(token, Settings.jwt_secret_key, algorithms=[Settings.jwt_algorithm])


def get_password_hash(password: str) -> str:
    return pwd_context.hash(password)


def verify_password(plain_password: str, hashed_password: str) -> bool:
    return pwd_context.verify(plain_password, hashed_password)
