backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
alembic==1.16.4
pydantic==2.11.7
pydantic-settings==2.10.1
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
python-multipart==0.0.20
redis==6.4.0
httpx==0.28.1
pytest==8.4.1
pytest-asyncio==1.1.0
websockets==13.1
ably==2.0.12
email-validator==2.2.0
